# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI-CAD Backend
  version: 0.1.0
  description: |
    Estonian-language prompt → parametric CAD spec → STL.
servers:
  - url: http://localhost:8080
  - url: https://cad.krerte.ee
paths:
  /api/templates:
    get:
      summary: List available parametric templates and their schemas
      responses:
        '200':
          description: catalog
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/TemplateSchema'
  /api/spec:
    post:
      summary: Turn an Estonian prompt into a structured spec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string, example: "vajan riiuliklambrit 32mm torule, 5kg koormus" }
      responses:
        '200':
          description: structured spec
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Spec' }
        '400':
          description: no template matched
  /api/generate:
    post:
      summary: Render STL for a (possibly user-tweaked) spec
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Spec' }
      responses:
        '200':
          description: STL binary
          content:
            application/sla:
              schema: { type: string, format: binary }
  /api/metrics:
    post:
      summary: Fast heuristic volume/bbox/rough-print-time (debounced slider feedback)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Spec' }
      responses:
        '200':
          description: heuristic metrics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Metrics' }
  /api/preview:
    post:
      summary: |
        Precise slicer-based print time, filament mass and EUR cost for a spec.
        Generates the STL internally and routes it through the PrusaSlicer
        sidecar. Falls back to the heuristic if the slicer is not configured
        or fails — the `source` field distinguishes the two.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Spec'
                - type: object
                  properties:
                    material:     { type: string, enum: [pla, petg] }
                    fillDensity:  { type: string, example: "20" }
                    layerHeight:  { type: string, example: "0.2" }
      responses:
        '200':
          description: precise (or fallback-heuristic) preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preview' }
  /api/review:
    post:
      summary: AI Design Review — Claude vision critique of the current design
      description: |
        Sends the user's original Estonian prompt, the resolved spec and a
        base64 PNG of the three.js preview to Claude. Claude replies (via a
        forced tool call) with a structured review: overall score, verdict,
        strengths, concerns and actionable suggestions. Suggestions that
        carry {param, new_value} can be auto-applied by the frontend.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec: { $ref: '#/components/schemas/Spec' }
                prompt_et:
                  type: string
                  description: The user's original Estonian request. Helps Claude judge intent-fit.
                  example: "riiuliklamber 32mm torule, 5kg koormus"
                image_base64:
                  type: string
                  description: |
                    Base64 PNG snapshot of the three.js viewer (no data: prefix).
                    Optional — Claude can review from spec alone, but vision makes it much better.
      responses:
        '200':
          description: structured review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '400':
          description: spec missing
        '502':
          description: Claude call failed
        '503':
          description: ANTHROPIC_API_KEY not configured
  /api/meshy:
    post:
      summary: Free-form mesh fallback via Meshy.ai
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
      responses:
        '200':
          description: GLB URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_url: { type: string, format: uri }
                  format:    { type: string, example: glb }
        '503':
          description: Meshy not configured
components:
  schemas:
    ParamSchema:
      type: object
      properties:
        type:    { type: string, example: number }
        unit:    { type: string, example: mm }
        min:     { type: number }
        max:     { type: number }
        default: { type: number }
    TemplateSchema:
      type: object
      properties:
        description: { type: string }
        params:
          type: object
          additionalProperties: { $ref: '#/components/schemas/ParamSchema' }
    Spec:
      type: object
      required: [template, params]
      properties:
        template:   { type: string, example: shelf_bracket }
        params:
          type: object
          additionalProperties: { type: number }
        summary_et: { type: string }
    Metrics:
      type: object
      properties:
        volume_cm3:               { type: number }
        bbox_mm:
          type: object
          properties:
            x: { type: number }
            y: { type: number }
            z: { type: number }
        weight_g_pla:             { type: number }
        print_time_min_estimate:  { type: integer }
        overhang_risk:            { type: boolean }
    Preview:
      type: object
      properties:
        source:              { type: string, enum: [slicer, heuristic, heuristic_slicer_failed] }
        preset:              { type: string, example: pla_default }
        print_time_sec:      { type: integer, example: 1397 }
        print_time_human:    { type: string, example: "23min 17s" }
        filament_length_m:   { type: number }
        filament_volume_cm3: { type: number }
        filament_g:          { type: number, example: 7.4 }
        filament_cost_eur:   { type: number, example: 0.19 }
        volume_cm3:          { type: number }
        bbox_mm:
          type: object
          properties:
            x: { type: number }
            y: { type: number }
            z: { type: number }
        overhang_risk:       { type: boolean }
    Suggestion:
      type: object
      required: [label_et, rationale_et]
      properties:
        label_et:     { type: string, example: "Suurenda seinapaksust 5mm peale" }
        rationale_et: { type: string, example: "Praegune 3mm võib 5kg all väänduma hakata." }
        param:
          type: string
          description: spec.params key to tweak (optional — non-numeric suggestions omit it)
          example: wall_thickness
        new_value:
          type: number
          description: New value for the param; UI clamps to template min/max.
          example: 5
    Review:
      type: object
      required: [score, verdict_et, strengths, weaknesses, suggestions]
      properties:
        score:      { type: integer, minimum: 1, maximum: 10, example: 7 }
        verdict_et: { type: string, example: "Kindel ja prinditav, aga ülakinnituse kõrgus võiks olla 2mm kõrgem." }
        strengths:
          type: array
          items: { type: string }
        weaknesses:
          type: array
          items: { type: string }
        suggestions:
          type: array
          items: { $ref: '#/components/schemas/Suggestion' }